Notebook 1 : Analyzing the student data
Commit includes a  Python notebook in .ipynb and .html formats and a csv
file containing the student data which is analysed to predict the
admissions to graduate school at UCLA. GRE scores, GPA scores and class
rank are used for the prediction.
A 2-layer Neural network is trained and used for the prediction.