Project 1 - Predict Bike Sharing Pattern
In this project, the task was to build a neural network and use it to
predict daily bike rental ridership. Some of the code was provided by
Udacity.
This commit contains the Python notebook in .ipynb and .html formats
which contains code for predicting bike sharing pattern. The Python file
"my_answers.py" contains code for initializng, training and testing the neural network.