Notebook 3e - Inference And Validation
Exercise was to perform Inference and validation using the trained neural network using PyTorch. 
Inference is the term used for making predictions using the trained network
If the neural network perform too well on training data and fails to generalize, it is called overfitting.  Validation is done to test for overfitting.
Some of the code was already provided by Udacity.